- Removed use of Python 3.7 function `datetime.fromisoformat()` - Added state object base class.
